fix Definitive Fusion string
When the activation condition that activated the effect of destroying the opponent's cards was met, a blank sign appeared with the condition "yes" or "no" without a description of the effect to be resolved. That sign did not show anything, when it should show "Destroy 1 of your opponent's cards?"

card ID: 71143015

There are a lot of errors like these in the strings, strings that do not appear, disordered strings or strings that indicate one thing and do another are some examples.